Finished receipt analyzer logic
